chore(Lemnisk): move sdk loader to different file
