+++ src/integrations/Lemnisk/browser.js
@@ -3,8 +3,8 @@
 /* eslint-disable func-names */
 /* eslint-disable class-methods-use-this */
 import Logger from '../../utils/logger';
-import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
 import { NAME } from './constants';
+import { loader } from './loader';
 
 const logger = new Logger(NAME);
 class Lemnisk {
@@ -23,34 +23,13 @@ class Lemnisk {
 
   init() {
     logger.debug('===in init Lemnisk Marketing Automation===');
-    (function (window, tag, o, a, r) {
-      var methods = ['init', 'page', 'track', 'identify'];
-      window.lmSMTObj = window.lmSMTObj || [];
-
-      for (var i = 0; i < methods.length; i++) {
-        lmSMTObj[methods[i]] = (function (methodName) {
-          return function () {
-            lmSMTObj.push([methodName].concat(Array.prototype.slice.call(arguments)));
-          };
-        })(methods[i]);
-      }
-      // eslint-disable-next-line no-param-reassign
-      a = o.getElementsByTagName('head')[0];
-      // eslint-disable-next-line no-param-reassign
-      r = o.createElement('script');
-      r.setAttribute('data-loader', LOAD_ORIGIN);
-      r.type = 'text/javascript';
-      r.async = 1;
-      r.src = tag;
-      a.appendChild(r);
-    })(
+    loader(
       window,
       document.location.protocol === 'https:'
         ? `https://cdn25.lemnisk.co/ssp/st/${this.accountId}.js`
         : `http://cdn25.lemnisk.co/ssp/st/${this.accountId}.js`,
       document,
     );
-    window.lmSMTObj.init(this.sdkWriteKey);
   }
 
   isLoaded() {
